clean

+++ app/controllers/tournaments_controller.rb
@@ -32,12 +32,18 @@ class TournamentsController < ApplicationController
     tournament = Tournament.find(params[:id])
     tournament.multiplier = permitted_params[:tournament][:multiplier]
     raw = params[:raw][:to_s]
-    if raw != tournament.to_raw
+    if raw != tournament.to_raw && raw.present?
       tournament.results.destroy_all
       Participant.where('id not in (select distinct(participant_id) FROM results)').map(&:destroy)
       tournament.construct_results(raw)
     end
-    if tournament.save
+    if tournament.valid?
+      if raw.blank?
+        Participant.where('id not in (select distinct(participant_id) FROM results)').map(&:destroy)
+        tournament.destroy
+      else
+        tournament.save
+      end
       redirect_to game
     else
       render :edit
